pyproject.toml
[tool.poetry]
name = "meridian-backend"
version = "0.1.0"
description = "Foundation phase analytics platform backend"
authors = ["Meridian Team <team@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
motor = "^3.3.1"
redis = "^5.0.3"
apscheduler = "^3.10.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
httpx = "^0.27.0"
prometheus-client = "^0.19.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-sdk = "^1.24.0"
sentry-sdk = "^1.40.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.3.5"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
coverage = "^7.4.3"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.mypy]
python_version = 3.11
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
